requirement.txt
asgiref==3.2.3
Django==3.0.3
django-redis==4.11.0
django-user-agents==0.4.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
PyJWT==1.7.1
pytz==2019.3
PyYAML==5.3
redis==3.4.1
six==1.14.0
sqlparse==0.3.0
ua-parser==0.9.0
user-agents==2.1
